r813: Integrated faster_nested_set plugin; changed file monitor to make a first pass at which only directories are added without files; changed daemon to be able to start from any directory; removed better_nested_set plugin; removed nested_set_improvements.rb; added on delete cascade constraints to file and directory schema, and a foreign key constraint to the directory parent column.

+++ daemon/file_monitor.rb
@@ -7,8 +7,19 @@ class FileMonitor
   def FileMonitor.run_on_new_directory(path, update_time)
     this_server = Earth::Server.this_server
     puts "WARNING: Watching new directory. So, clearing out database"
+
+    # Workaround until on delete cascade is properly supported by faster_nested_set.  -->
+    # FIXME: The following line unconditionally uses PostgreSQL extensions
+    # Provide an alternative which uses a sub-select
+    # FIXME: Do this in a transaction
+    Earth::File.connection.delete("DELETE FROM #{Earth::File.table_name} USING #{Earth::Directory.table_name} WHERE #{Earth::Directory.table_name}.server_id=#{Earth::Server.this_server.id} AND #{Earth::Directory.table_name}.id = #{Earth::File.table_name}.directory_id", "Earth::Directory Delete all")
+    Earth::Directory.connection.delete("DELETE FROM #{Earth::Directory.table_name} WHERE server_id=#{Earth::Server.this_server.id}", "Earth::Directory Delete all")
+    # <--
+    
     this_server.directories.clear      
     directory = this_server.directories.create(:name => File.expand_path(path))
+    update(directory, :only_build_directories => true)
+    directory.save
     run(directory, update_time)
   end
   
@@ -34,13 +45,14 @@ private
     end
   end
   
-  def FileMonitor.update(directory)
+  def FileMonitor.update(directory, only_build_directories = false)
     directory.each do |d|
-      update_non_recursive(d)
+      update_non_recursive(d, only_build_directories)
     end
   end
 
-  def FileMonitor.update_non_recursive(directory)
+  def FileMonitor.update_non_recursive(directory, only_build_directories = false)
+
     # TODO: remove exist? call as it is an extra filesystem access
     if File.exist?(directory.path)
       new_directory_stat = File.lstat(directory.path)
@@ -61,37 +73,44 @@ private
     added_directory_names = subdirectory_names - directory.children.map{|x| x.name}
     added_directory_names.each do |name|
       dir = Earth::Directory.benchmark("Creating directory with name #{directory.name}", Logger::DEBUG, !log_all_sql) do
-        directory.child_create(:name => name, :server_id => directory.server_id)
+        attributes = { :name => name, :server_id => directory.server_id }
+        if only_build_directories then
+          directory.children.build(attributes)
+        else
+          directory.children.create(attributes)
+        end
       end
-      update_non_recursive(dir)
+      update_non_recursive(dir, only_build_directories)
     end
 
-    # By adding and removing files on the association, the cache of the association will be kept up to date
-    added_file_names = file_names - directory.files.map{|x| x.name}
-    added_file_names.each do |name|
-      Earth::File.benchmark("Creating file with name #{name}", Logger::DEBUG, !log_all_sql) do
-        directory.files.create(:name => name, :stat => stats[name])
+    if not only_build_directories then
+      # By adding and removing files on the association, the cache of the association will be kept up to date
+      added_file_names = file_names - directory.files.map{|x| x.name}
+      added_file_names.each do |name|
+        Earth::File.benchmark("Creating file with name #{name}", Logger::DEBUG, !log_all_sql) do
+          directory.files.create(:name => name, :stat => stats[name])
+        end
       end
-    end
 
-    directory.files.each do |file|
-      # If the file still exists
-      if file_names.include?(file.name)
-        # If the file has changed
-        if file.stat != stats[file.name]
-          file.stat = stats[file.name]
-          Earth::File.benchmark("Updating file with name #{file.name}", Logger::DEBUG, !log_all_sql) do
-            file.save
+      directory.files.each do |file|
+        # If the file still exists
+        if file_names.include?(file.name)
+          # If the file has changed
+          if file.stat != stats[file.name]
+            file.stat = stats[file.name]
+            Earth::File.benchmark("Updating file with name #{file.name}", Logger::DEBUG, !log_all_sql) do
+              file.save
+            end
+          end
+          # If the file has been deleted
+        else
+          Earth::Directory.benchmark("Removing file with name #{file.name}", Logger::DEBUG, !log_all_sql) do
+            directory.files.delete(file)
           end
         end
-      # If the file has been deleted
-      else
-        Earth::Directory.benchmark("Removing file with name #{file.name}", Logger::DEBUG, !log_all_sql) do
-          directory.files.delete(file)
-        end
       end
     end
-    
+
     directory.children.each do |dir|
       # If the directory has been deleted
       if !subdirectory_names.include?(dir.name)
@@ -102,11 +121,13 @@ private
     end
     
     # Update the directory stat information at the end
-    if File.exist?(directory.path)
-      directory.stat = new_directory_stat
-      # This will not overwrite 'lft' and 'rgt' so it doesn't matter if these are out of date
-      Earth::Directory.benchmark("Updating directory with name #{directory.name}", Logger::DEBUG, !log_all_sql) do
-        directory.update
+    if not only_build_directories then
+      if File.exist?(directory.path)
+        directory.stat = new_directory_stat
+        # This will not overwrite 'lft' and 'rgt' so it doesn't matter if these are out of date
+        Earth::Directory.benchmark("Updating directory with name #{directory.name}", Logger::DEBUG, !log_all_sql) do
+          directory.update
+        end
       end
     end
   end
